Brain/backend/requirements.txt
flask
flask-cors
torch
nibabel
numpy
monai
pillow
matplotlib
